+++ UserDataDownloadOrchestrator/__tests__/handler.test.ts
@@ -1,32 +1,65 @@
 // tslint:disable: no-any
 
 import { IFunctionContext } from "durable-functions/lib/src/classes";
-import { UserDataProcessingChoiceEnum } from "io-functions-commons/dist/generated/definitions/UserDataProcessingChoice";
 import { UserDataProcessingStatusEnum } from "io-functions-commons/dist/generated/definitions/UserDataProcessingStatus";
-import { UserDataProcessing } from "io-functions-commons/dist/src/models/user_data_processing";
 import {
-  mockCallSubOrchestrator,
+  mockOrchestratorCallActivity,
+  mockOrchestratorCallActivityWithRetry,
   mockOrchestratorContext,
   mockOrchestratorGetInput
 } from "../../__mocks__/durable-functions";
-import { aUserDataProcessing } from "../../__mocks__/mocks";
-import { handler } from "../handler";
+import { aArchiveInfo, aUserDataProcessing } from "../../__mocks__/mocks";
+import { ActivityResultSuccess as ExtractUserDataActivityResultSuccess } from "../../ExtractUserDataActivity/handler";
+import { ActivityResultSuccess as SendUserDataDownloadMessageActivityResultSuccess } from "../../SendUserDataDownloadMessageActivity/handler";
+import {
+  ActivityFailure,
+  getHandler,
+  InvalidInputFailure,
+  OrchestratorSuccess
+} from "../handler";
 
-const aProcessableDocument = {
-  ...aUserDataProcessing,
-  choice: UserDataProcessingChoiceEnum.DOWNLOAD,
-  status: UserDataProcessingStatusEnum.PENDING
-};
+import { Millisecond } from "italia-ts-commons/lib/units";
+import { ActivityResultSuccess as SetUserDataProcessingStatusActivityResultSuccess } from "../../SetUserDataProcessingStatusActivity/handler";
 
-const aNonProcessableDocumentWrongStatus = {
-  ...aUserDataProcessing,
-  status: UserDataProcessingStatusEnum.WIP
-};
+const DELAY = 1 as Millisecond;
 
-const aNonProcessableDocumentWrongChoice = {
-  ...aUserDataProcessing,
-  choice: UserDataProcessingChoiceEnum.DELETE
-};
+const aNonSuccess = "any non-success value";
+
+const setUserDataProcessingStatusActivity = jest.fn().mockImplementation(() =>
+  SetUserDataProcessingStatusActivityResultSuccess.encode({
+    kind: "SUCCESS",
+    value: aUserDataProcessing
+  })
+);
+const extractUserDataActivity = jest.fn().mockImplementation(() =>
+  ExtractUserDataActivityResultSuccess.encode({
+    kind: "SUCCESS",
+    value: aArchiveInfo
+  })
+);
+const sendUserDataDownloadMessageActivity = jest
+  .fn()
+  .mockImplementation(() =>
+    SendUserDataDownloadMessageActivityResultSuccess.encode({ kind: "SUCCESS" })
+  );
+
+// A mock implementation proxy for df.callActivity/df.df.callActivityWithRetry that routes each call to the correct mock implentation
+const switchMockImplementation = (name: string, ...args: readonly unknown[]) =>
+  (name === "SetUserDataProcessingStatusActivity"
+    ? setUserDataProcessingStatusActivity
+    : name === "ExtractUserDataActivity"
+    ? extractUserDataActivity
+    : name === "SendUserDataDownloadMessageActivity"
+    ? sendUserDataDownloadMessageActivity
+    : jest.fn())(name, ...args);
+
+// I assign switchMockImplementation to both because
+// I don't want tests to depend on implementation details
+// such as which activity is called with retry and which is not
+mockOrchestratorCallActivity.mockImplementation(switchMockImplementation);
+mockOrchestratorCallActivityWithRetry.mockImplementation(
+  switchMockImplementation
+);
 
 /**
  * Util function that takes an orchestrator and executes each step until is done
@@ -49,40 +82,225 @@ const consumeOrchestrator = (orch: any) => {
 // just a convenient cast, good for every test case
 const context = (mockOrchestratorContext as unknown) as IFunctionContext;
 
+const handler = getHandler(DELAY);
+
+// tslint:disable-next-line: no-big-function
 describe("handler", () => {
   beforeEach(() => {
     jest.clearAllMocks();
   });
 
   it("should fail on invalid input", () => {
-    const input = "invalid";
-    mockOrchestratorGetInput.mockReturnValueOnce(input);
+    const document = "invalid";
+    mockOrchestratorGetInput.mockReturnValueOnce(document);
 
-    try {
-      consumeOrchestrator(handler(context));
-      fail("it should throw");
-    } catch (error) {
-      expect(mockCallSubOrchestrator).not.toHaveBeenCalled();
-    }
+    const result = consumeOrchestrator(handler(context));
+
+    expect(InvalidInputFailure.decode(result).isRight()).toBe(true);
+    expect(setUserDataProcessingStatusActivity).not.toHaveBeenCalled();
+    expect(extractUserDataActivity).not.toHaveBeenCalled();
+    expect(sendUserDataDownloadMessageActivity).not.toHaveBeenCalled();
   });
 
-  it("should process every processable document", () => {
-    const processableDocs: ReadonlyArray<UserDataProcessing> = [
-      aProcessableDocument,
-      aProcessableDocument,
-      aProcessableDocument
-    ];
+  it.each`
+    name        | status
+    ${"WIP"}    | ${UserDataProcessingStatusEnum.WIP}
+    ${"CLOSED"} | ${UserDataProcessingStatusEnum.CLOSED}
+  `("should skip if the status is $name", ({ status }) => {
+    const document = {
+      ...aUserDataProcessing,
+      status
+    };
+    mockOrchestratorGetInput.mockReturnValueOnce(document);
 
-    const input: ReadonlyArray<any> = [
-      ...processableDocs,
-      aNonProcessableDocumentWrongStatus,
-      aNonProcessableDocumentWrongChoice
-    ];
-    mockOrchestratorGetInput.mockReturnValueOnce(input);
+    const result = consumeOrchestrator(handler(context));
 
-    consumeOrchestrator(handler(context));
-    expect(mockCallSubOrchestrator).toHaveBeenCalledTimes(
-      processableDocs.length
+    expect(InvalidInputFailure.decode(result).isRight()).toBe(true);
+    expect(setUserDataProcessingStatusActivity).not.toHaveBeenCalled();
+    expect(extractUserDataActivity).not.toHaveBeenCalled();
+    expect(sendUserDataDownloadMessageActivity).not.toHaveBeenCalled();
+  });
+
+  it("should success if everything goes well", () => {
+    const document = {
+      ...aUserDataProcessing,
+      status: UserDataProcessingStatusEnum.PENDING
+    };
+    mockOrchestratorGetInput.mockReturnValueOnce(document);
+
+    const result = consumeOrchestrator(handler(context));
+
+    expect(OrchestratorSuccess.decode(result).isRight()).toBe(true);
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalledTimes(2);
+    // first, set as WIP
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalledWith(
+      expect.any(String),
+      {
+        currentRecord: expect.any(Object),
+        nextStatus: UserDataProcessingStatusEnum.WIP
+      }
     );
+    // then, set as CLOSED
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalledWith(
+      expect.any(String),
+      {
+        currentRecord: expect.any(Object),
+        nextStatus: UserDataProcessingStatusEnum.CLOSED
+      }
+    );
+    expect(extractUserDataActivity).toHaveBeenCalled();
+    expect(sendUserDataDownloadMessageActivity).toHaveBeenCalled();
+  });
+
+  it("should set as FAILED when data extraction fails", () => {
+    extractUserDataActivity.mockImplementationOnce(() => aNonSuccess);
+
+    const document = {
+      ...aUserDataProcessing,
+      status: UserDataProcessingStatusEnum.PENDING
+    };
+    mockOrchestratorGetInput.mockReturnValueOnce(document);
+
+    const result = consumeOrchestrator(handler(context));
+
+    expect(ActivityFailure.decode(result).isRight()).toBe(true);
+    expect(result.activityName).toBe("ExtractUserDataActivity");
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalled(); // any times, at least one
+    // then, set as FAILED
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalledWith(
+      expect.any(String),
+      {
+        currentRecord: expect.any(Object),
+        nextStatus: UserDataProcessingStatusEnum.FAILED
+      }
+    );
+    expect(extractUserDataActivity).toHaveBeenCalled();
+    expect(sendUserDataDownloadMessageActivity).not.toHaveBeenCalled();
+  });
+
+  it("should set as FAILED when send message fails", () => {
+    sendUserDataDownloadMessageActivity.mockImplementationOnce(
+      () => aNonSuccess
+    );
+
+    const document = {
+      ...aUserDataProcessing,
+      status: UserDataProcessingStatusEnum.PENDING
+    };
+    mockOrchestratorGetInput.mockReturnValueOnce(document);
+
+    const result = consumeOrchestrator(handler(context));
+
+    expect(ActivityFailure.decode(result).isRight()).toBe(true);
+    expect(result.activityName).toBe("SendUserDataDownloadMessageActivity");
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalled(); // any times, at least one
+    // then, set as FAILED
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalledWith(
+      expect.any(String),
+      {
+        currentRecord: expect.any(Object),
+        nextStatus: UserDataProcessingStatusEnum.FAILED
+      }
+    );
+    expect(extractUserDataActivity).toHaveBeenCalled();
+    expect(sendUserDataDownloadMessageActivity).toHaveBeenCalled();
+  });
+
+  it("should set as FAILED when status update to WIP fails", () => {
+    setUserDataProcessingStatusActivity.mockImplementationOnce(
+      () => aNonSuccess
+    );
+
+    const document = {
+      ...aUserDataProcessing,
+      status: UserDataProcessingStatusEnum.PENDING
+    };
+    mockOrchestratorGetInput.mockReturnValueOnce(document);
+
+    const result = consumeOrchestrator(handler(context));
+
+    expect(ActivityFailure.decode(result).isRight()).toBe(true);
+    expect(result.activityName).toBe("SetUserDataProcessingStatusActivity");
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalled(); // any times, at least one
+    // then, set as FAILED
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalledWith(
+      expect.any(String),
+      {
+        currentRecord: expect.any(Object),
+        nextStatus: UserDataProcessingStatusEnum.FAILED
+      }
+    );
+    expect(extractUserDataActivity).not.toHaveBeenCalled();
+    expect(sendUserDataDownloadMessageActivity).not.toHaveBeenCalled();
+  });
+
+  it("should set as FAILED when status update to WIP fails", () => {
+    setUserDataProcessingStatusActivity.mockImplementationOnce(
+      () => aNonSuccess
+    );
+
+    const document = {
+      ...aUserDataProcessing,
+      status: UserDataProcessingStatusEnum.PENDING
+    };
+    mockOrchestratorGetInput.mockReturnValueOnce(document);
+
+    const result = consumeOrchestrator(handler(context));
+
+    expect(ActivityFailure.decode(result).isRight()).toBe(true);
+    expect(result.activityName).toBe("SetUserDataProcessingStatusActivity");
+    expect(result.extra).toEqual({
+      status: UserDataProcessingStatusEnum.WIP
+    });
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalled(); // any times, at least one
+    // then, set as FAILED
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalledWith(
+      expect.any(String),
+      {
+        currentRecord: expect.any(Object),
+        nextStatus: UserDataProcessingStatusEnum.FAILED
+      }
+    );
+    expect(extractUserDataActivity).not.toHaveBeenCalled();
+    expect(sendUserDataDownloadMessageActivity).not.toHaveBeenCalled();
+  });
+
+  it("should set as FAILED when status update to CLOSED fails", () => {
+    // the first time is called is for WIP
+    setUserDataProcessingStatusActivity.mockImplementationOnce(() =>
+      SetUserDataProcessingStatusActivityResultSuccess.encode({
+        kind: "SUCCESS",
+        value: aUserDataProcessing
+      })
+    );
+
+    setUserDataProcessingStatusActivity.mockImplementationOnce(
+      () => aNonSuccess
+    );
+
+    const document = {
+      ...aUserDataProcessing,
+      status: UserDataProcessingStatusEnum.PENDING
+    };
+    mockOrchestratorGetInput.mockReturnValueOnce(document);
+
+    const result = consumeOrchestrator(handler(context));
+
+    expect(ActivityFailure.decode(result).isRight()).toBe(true);
+    expect(result.activityName).toBe("SetUserDataProcessingStatusActivity");
+    expect(result.extra).toEqual({
+      status: UserDataProcessingStatusEnum.CLOSED
+    });
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalled(); // any times, at least one
+    // then, set as FAILED
+    expect(setUserDataProcessingStatusActivity).toHaveBeenCalledWith(
+      expect.any(String),
+      {
+        currentRecord: expect.any(Object),
+        nextStatus: UserDataProcessingStatusEnum.FAILED
+      }
+    );
+    expect(extractUserDataActivity).toHaveBeenCalled();
+    expect(sendUserDataDownloadMessageActivity).toHaveBeenCalled();
   });
 });
