+++ UserDataDownloadOrchestrator/function.json
@@ -6,5 +6,5 @@
         "direction": "in"
       }
     ],
-    "scriptFile": "../dist/UserDataDownloadOrchestrator/index.js"
+    "scriptFile": "../dist/UserDataDownloadSubOrchestrator/index.js"
 }
