--- UserDataDownloadOrchestratorTrigger/index.ts
@@ -1,7 +1,0 @@
-import { Context } from "@azure/functions";
-import * as df from "durable-functions";
-
-export default (context: Context, input: unknown) =>
-  df
-    .getClient(context)
-    .startNew("UserDataDownloadOrchestrator", undefined, input);
