--- UserDataDownloadSubOrchestrator/function.json
@@ -1,10 +1,0 @@
-{
-    "bindings": [
-      {
-        "name": "context",
-        "type": "orchestrationTrigger",
-        "direction": "in"
-      }
-    ],
-    "scriptFile": "../dist/UserDataDownloadSubOrchestrator/index.js"
-}
