+++ UserDataProcessingTrigger/__tests__/index.test.ts
@@ -1,0 +1,59 @@
+// tslint:disable: no-any
+
+import { UserDataProcessingChoiceEnum } from "io-functions-commons/dist/generated/definitions/UserDataProcessingChoice";
+import { UserDataProcessingStatusEnum } from "io-functions-commons/dist/generated/definitions/UserDataProcessingStatus";
+import { UserDataProcessing } from "io-functions-commons/dist/src/models/user_data_processing";
+import { context, mockStartNew } from "../../__mocks__/durable-functions";
+import { aUserDataProcessing } from "../../__mocks__/mocks";
+import handler from "../index";
+
+const aProcessableDownload = {
+  ...aUserDataProcessing,
+  choice: UserDataProcessingChoiceEnum.DOWNLOAD,
+  status: UserDataProcessingStatusEnum.PENDING
+};
+
+const aNonProcessableDownloadWrongStatus = {
+  ...aUserDataProcessing,
+  status: UserDataProcessingStatusEnum.WIP
+};
+
+const aNonProcessableDownloadWrongChoice = {
+  ...aUserDataProcessing,
+  choice: UserDataProcessingChoiceEnum.DELETE
+};
+
+describe("UserDataProcessingTrigger", () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it("should fail on invalid input", () => {
+    const input = "invalid";
+
+    try {
+      handler(context, input);
+      fail("it should throw");
+    } catch (error) {
+      expect(mockStartNew).not.toHaveBeenCalled();
+    }
+  });
+
+  it("should process every processable document", () => {
+    const processableDocs: ReadonlyArray<UserDataProcessing> = [
+      aProcessableDownload,
+      aProcessableDownload,
+      aProcessableDownload
+    ];
+
+    const input: ReadonlyArray<any> = [
+      ...processableDocs,
+      aNonProcessableDownloadWrongStatus,
+      aNonProcessableDownloadWrongChoice
+    ];
+
+    handler(context, input);
+
+    expect(mockStartNew).toHaveBeenCalledTimes(processableDocs.length);
+  });
+});
